map whole config directory
The example docker run command maps the local volume to the lib subdirectory, causing the machine configuration files to not be backed up on the host. I suggest removing lib/.

Note that when mapping the configuration files, the directory and file permissions must match uid/gid 1000 on the host for the container to be able to write the files. docker logs anisette-v3 will indicate such problems if the container doesn't start.